BugzID:9592 Also BugzID:9581 Change to MarkerRPCImpl class..wasnt chekcing if returned alias id was null before deleting its attribution
git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@28793 908b024a-c82b-635c-c293-f2b15b23277c

+++ source/org/zfin/gwt/root/server/MarkerRPCServiceImpl.java
@@ -134,7 +134,9 @@ public class MarkerRPCServiceImpl extends ZfinRemoteServiceServlet implements Ma
         HibernateUtil.createTransaction();
         infrastructureRepository.deleteRecordAttribution(markerZdbID, pubZdbID);
         DataAlias dAlias=infrastructureRepository.getDataAliasByID(markerZdbID);
+        if (dAlias!=null) {
         infrastructureRepository.deleteRecordAttribution(dAlias.getZdbID(), pubZdbID);
+        }
 
         infrastructureRepository.insertUpdatesTable(markerZdbID, "record attribution", pubZdbID, "removed", "Removed direct attribution");
         HibernateUtil.flushAndCommitCurrentSession();
